refactor: team, place 상수 분리
