+++ src/components/chip/season-chip/index.tsx
@@ -1,7 +1,7 @@
 import { useRecoilValue } from "recoil";
 import * as S from "./SeasonChip.style";
 import { currSeasonState, teamState } from "../../../recoil/system";
-import { TEAM_COLOR } from "../../../constants/team";
+import { TEAM_DATA } from "../../../constants/team";
 
 interface SeasonChipProps {
   season: number;
@@ -16,7 +16,9 @@ const SeasonChip = ({ season, onClick }: SeasonChipProps) => {
     <S.Container
       isActive={currSeason === season}
       team={
-        currTeam.length > 0 ? TEAM_COLOR[currTeam] : "var(--color-gray-200)"
+        currTeam.length > 0
+          ? TEAM_DATA[currTeam].color
+          : "var(--color-gray-200)"
       }
       onClick={onClick}
     >
